break out unicode packaage
