+++ unicode/emoji.go
@@ -1,4 +1,4 @@
-package log
+package unicode
 
 import (
 	"unicode"
